requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-Login==0.6.3
python-dotenv==1.0.0
Werkzeug==3.0.1
eventlet==0.35.1
